requirements.txt
cinemagoer==2022.2.11
greenlet==1.1.2
install==1.3.5
lxml==4.8.0
PyYAML==6.0
SQLAlchemy==1.4.36
